# #25012: Add int32 support for argmax