--- a/tt_metal/hw/inc/utils/float32.h
+++ b/tt_metal/hw/inc/utils/float32.h
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: © 2043 Tenstorrent Inc.
+// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 //
 // SPDX-License-Identifier: Apache-2.0