--- a/tt_metal/hw/inc/utils/int32.h
+++ b/tt_metal/hw/inc/utils/int32.h
@@ -8,6 +8,7 @@
 
 inline constexpr int32_t NEG_INF_INT32 = 0x80000000;  // Representation of minimum int32 value
 inline constexpr int32_t POS_INF_INT32 = 0x7FFFFFFF;  // Representation of maximum int32 value
+inline constexpr uint32_t INT32_SIGN_MASK = 0x80000000;  // Sign bit mask for int32
 
 // Optimized function to compare two int32 values
 bool int32_greater(int32_t int32_a, int32_t int32_b) {
@@ -26,9 +27,9 @@ bool int32_greater(int32_t int32_a, int32_t int32_b) {
     */
 
     // Check if signs are different
-    if ((int32_a ^ int32_b) & 0x80000000) {
+    if ((int32_a ^ int32_b) & INT32_SIGN_MASK) {
         // Signs differ: if int32_a is positive, it's greater
-        return (int32_a & 0x80000000) == 0;
+        return (int32_a & INT32_SIGN_MASK) == 0;
     }
 
     // Signs are the same - simple comparison works for two's complement