--- a/ttnn/cpp/ttnn/operations/reduction/argmax/device/argmax_op.cpp
+++ b/ttnn/cpp/ttnn/operations/reduction/argmax/device/argmax_op.cpp
@@ -117,10 +117,10 @@ operation::ProgramWithCallbacks ArgMax::create_program(
     const auto& output_tensor = output_tensors.at(0);
     const auto normalized_dim =
         this->dim.has_value() ? *this->dim + input_tensor.padded_shape().rank() * (*this->dim < 0) : this->dim;
-    // if (this->use_multicore) {
-    //     return detail::argmax_multi_core(
-    //         input_tensor, output_tensor, normalized_dim, this->keepdim, this->sub_core_grids);
-    // }
+    if (this->use_multicore) {
+        return detail::argmax_multi_core(
+            input_tensor, output_tensor, normalized_dim, this->keepdim, this->sub_core_grids);
+    }
     return detail::argmax_single_core(input_tensor, output_tensor, normalized_dim, this->keepdim);
 }