--- a/ttnn/cpp/ttnn/operations/reduction/argmax/device/kernels/argmax_common.hpp
+++ b/ttnn/cpp/ttnn/operations/reduction/argmax/device/kernels/argmax_common.hpp
@@ -16,7 +16,7 @@ auto get_default_value() {
     if constexpr (data_format == DataFormat::Float16_b) {
         return uint16_t{NEG_INF_BFLOAT16};
     } else if constexpr (data_format == DataFormat::Float32) {
-        return float{NEG_INF_FLOAT32};
+        return uint32_t{NEG_INF_FLOAT32};
     } else if constexpr (data_format == DataFormat::Int32) {
         return int32_t{NEG_INF_INT32};
     } else if constexpr (data_format == DataFormat::UInt32) {
@@ -46,8 +46,8 @@ void compare_values(
             max_val = val;
         }
     } else if constexpr (data_format == DataFormat::Float32) {
-        volatile tt_l1_ptr float* in_vals = reinterpret_cast<volatile tt_l1_ptr float*>(src_cb_addr);
-        float val = in_vals[i];
+        volatile tt_l1_ptr uint32_t* in_vals = reinterpret_cast<volatile tt_l1_ptr uint32_t*>(src_cb_addr);
+        uint32_t val = in_vals[i];
         if (float32_greater(val, max_val)) {
             max_idx = reduce_all ? (k * inner_dim_units * red_dim_units + j * red_dim_units + i) : i;
             max_val = val;