… tests/sweep_framework/sweeps/reduction/argmax/argmax.py
… tests/ttnn/unit_tests/operations/reduce/test_argmax.py
… tt_metal/hw/CMakeLists.txt
… ttnn/cpp/ttnn/operations/reduction/argmax/device/kernels/reader_argmax_interleaved.cpp