### Ticket
https://github.com/tenstorrent/tt-metal/issues/25012

### Problem description
TTNN Argmax OP supported only BFloat16.

### What's changed
- Added support for Float32, UInt32, Int32 and UInt16,
- Updated docs,
- Added tests for different datatypes.

### Check that performance did not change

Tiles | Type | DEVICE KERNEL DURATION PER CORE MIN [ms] | DEVICE KERNEL DURATION PER CORE MAX [ms] | DEVICE KERNEL DURATION PER CORE AVG [ms]
-- | -- | -- | -- | --
Current implementation
64 | Single Core | 0.054 | 0.054 | 0.054
128 | Single Core | 2.856 | 2.856 | 2.856
256 | Single Core | 5.696 | 5.696 | 5.696
512 | Single Core | 13.487 | 13.487 | 13.487
1024 | Single Core | 26.961 | 26.961 | 26.961
2048 | Single Core | 44.491 | 44.491 | 44.491
4096 | Single Core | 107.798 | 107.798 | 107.798
64 | Mult Core | 0.002 | 0.005 | 0.002
128 | Mult Core | 0.078 | 0.147 | 0.080
256 | Mult Core | 0.143 | 0.213 | 0.146
512 | Mult Core | 0.273 | 0.345 | 0.277
1024 | Mult Core | 0.533 | 0.608 | 0.538
2048 | Mult Core | 1.053 | 1.132 | 1.060
4096 | Mult Core | 2.092 | 2.184 | 2.106
New implementation
64 | Single Core | 0.054 | 0.054 | 0.054
128 | Single Core | 2.855 | 2.855 | 2.855
256 | Single Core | 5.696 | 5.696 | 5.696
512 | Single Core | 13.487 | 13.487 | 13.487
1024 | Single Core | 26.961 | 26.961 | 26.961
2048 | Single Core | 44.491 | 44.491 | 44.491
4096 | Single Core | 107.798 | 107.798 | 107.798
64 | Mult Core | 0.002 | 0.005 | 0.002
128 | Mult Core | 0.078 | 0.147 | 0.080
256 | Mult Core | 0.143 | 0.213 | 0.146
512 | Mult Core | 0.273 | 0.345 | 0.277
1024 | Mult Core | 0.533 | 0.608 | 0.538
2048 | Mult Core | 1.052 | 1.132 | 1.061
4096 | Mult Core | 2.092 | 2.184 | 2.106

### Checklist
- [x] [All post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml) CI passes https://github.com/tenstorrent/tt-metal/actions/runs/16477071480
- [x] [Blackhole Post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml) CI with demo tests passes (if applicable) https://github.com/tenstorrent/tt-metal/actions/runs/16444921453
- [ ] [Model regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml) CI passes (if applicable)
- [ ] [Device performance regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml) CI passes (if applicable)
- [ ] (For models and ops writers) [Single-card demo tests](https://github.com/tenstorrent/tt-metal/actions/workflows/single-card-demo-tests.yaml) CI passes (if applicable) See [recommended dev flow](https://github.com/tenstorrent/tt-metal/blob/main/models/MODEL_ADD.md#a-recommended-dev-flow-on-github-for-adding-new-models).
- [ ] [Galaxy quick](https://github.com/tenstorrent/tt-metal/actions/workflows/tg-quick-trigger.yaml) CI passes (if applicable)
- [ ] [TG demo tests, for Llama](https://github.com/tenstorrent/tt-metal/actions/workflows/tg-demo-tests.yaml) CI passes, if applicable, because of current Llama work
- [ ] (For runtime and ops writers) [T3000 unit tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-unit-tests.yaml) CI passes (if applicable, since this is run on push to main)
- [ ] (For models and ops writers) [T3000 demo tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-demo-tests.yaml) CI passes (if applicable, since this is required for release)
- [ ] (For models and ops writers) [Single-card demo tests](https://github.com/tenstorrent/tt-metal/actions/workflows/single-card-demo-tests.yaml) CI passes (if applicable, since this is required for release)
- [x] New/Existing tests provide coverage for changes